remove Namespace and separate out DataSourceIcon
