+++ public/app/features/alerting/unified/rule-list/components/DataSourceSection.tsx
@@ -12,7 +12,7 @@ import { Spacer } from '../../components/Spacer';
 import { WithReturnButton } from '../../components/WithReturnButton';
 import { isAdmin, stringifyErrorLike } from '../../utils/misc';
 
-import { DataSourceIcon } from './Namespace';
+import { DataSourceIcon } from './DataSourceIcon';
 import { LoadingIndicator } from './RuleGroup';
 
 export interface DataSourceSectionProps extends PropsWithChildren {
